Created getting real date time for a history and an archived history

+++ app/src/main/java/com/cobaltumapps/simplecalculator/v15/calculator/services/room/model/ArchivedHistory.kt
@@ -11,6 +11,6 @@ data class ArchivedHistory(
     var id: Int?,
     var user_expression: String,
     var result_calculation: String,
-    var date_time_calculation: Int,
-    var date_time_archived: Int
+    var date_time_calculation: Long,
+    var date_time_archived: Long
 ): Serializable
